extract css variable utils
